Moved JS/Python tests.
Moved the javascript test with the rest of the javascript code.
Move the python tests out of the namespaced "python" part of the
test directory.
