+++ brew_view/static/test/e2e/scenarios.js
