+++ brew_view/static/test/karma-conf.js
