+++ brew_view/static/test/unit/controllers/about_test.js
