+++ brew_view/static/test/unit/controllers/admin_queue_test.js
