+++ brew_view/static/test/unit/controllers/application_test.js
