+++ brew_view/static/test/unit/controllers/command_index_test.js
