+++ brew_view/static/test/unit/controllers/command_view_test.js
