+++ brew_view/static/test/unit/controllers/request_view_test.js
