+++ brew_view/static/test/unit/controllers/system_view_test.js
