+++ brew_view/static/test/unit/directives/loading_test.js
