+++ brew_view/static/test/unit/directives/server_error_test.js
