+++ brew_view/static/test/unit/services/command_service_test.js
