+++ brew_view/static/test/unit/services/queue_service_test.js
