+++ brew_view/static/test/unit/services/version_service_test.js
