+++ test/unit/controllers/command_list_api_test.py
